refactor(components/react/loginForm): :truck: Se renombra use-form-options a register-options, pues son opciones del registro
