+++ src/components/react/LoginForm/utils/register-options/email.js
