+++ src/components/react/LoginForm/utils/register-options/index.js
