+++ src/components/react/LoginForm/utils/register-options/password.js
